Fix clippy -D warnings across entire codebase
Applied cargo clippy --fix to resolve all 93+ clippy errors that were
blocking the strict -D warnings gate. Changes span 96 files and include:
- Derivable Default impls
- Unnecessary or/else branches simplified
- Redundant closures replaced with method references
- Field reassignment patterns cleaned up
- Various idiomatic Rust improvements

All 563 lib tests pass. cargo clippy --all-targets -D warnings exits 0.

+++ src/server/api/web_fetch.rs
@@ -169,14 +169,14 @@ async fn execute_combo_fetch(
 ) -> Result<Response, crate::core::combo::ComboExecutionError> {
     let url = url.clone();
     let format = format.to_string();
-    let max_chars = max_chars as usize;
+    let max_chars = max_chars;
     let state = state.clone();
 
     crate::core::combo::execute_combo_strategy(models, combo_name, strategy, move |model: &str| {
         let model_owned = model.to_string();
         let url = url.clone();
         let format = format.clone();
-        let max_chars = max_chars as usize;
+        let max_chars = max_chars;
         let state = state.clone();
         async move {
             execute_single_fetch(&state, &model_owned, &url, &format, max_chars)
